데이터 변경

+++ Assets/_Project/Scripts/03_Data/Spec/SpecDatas.cs
@@ -26,9 +26,9 @@ namespace Pxp.Data
         /// 자동 획득 시간(분)
         public int auto_acquire_time;
         /// 자동 획득시 최대 한도
-        public double auto_max_count;
+        public int auto_max_count;
         /// 기본값
-        public double init_count;
+        public int init_count;
         /// 아이콘 키
         public string icon_key;
     }
